chore: move spinner and page loading component
