+++ apps/web/src/components/Messages/Composer.tsx
@@ -1,4 +1,3 @@
-import { Spinner } from '@components/UI/Spinner';
 import useWindowSize from '@components/utils/hooks/useWindowSize';
 import { ArrowRightIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
@@ -9,7 +8,7 @@ import { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
 import { useMessagePersistStore } from 'src/store/message';
 import { MESSAGES } from 'src/tracking';
-import { Button, Input } from 'ui';
+import { Button, Input, Spinner } from 'ui';
 
 interface ComposerProps {
   sendMessage: (message: string) => Promise<boolean>;
