+++ apps/web/src/components/Publication/Actions/Mirror.tsx
@@ -1,5 +1,4 @@
 import type { ApolloCache } from '@apollo/client';
-import { Spinner } from '@components/UI/Spinner';
 import { Tooltip } from '@components/UI/Tooltip';
 import { SwitchHorizontalIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
@@ -21,6 +20,7 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { PUBLICATION } from 'src/tracking';
+import { Spinner } from 'ui';
 import getSignature from 'utils/getSignature';
 import humanize from 'utils/humanize';
 import { publicationKeyFields } from 'utils/keyFields';
