+++ apps/web/src/components/Settings/Account/SuperFollow.tsx
@@ -1,4 +1,3 @@
-import { Spinner } from '@components/UI/Spinner';
 import { StarIcon, XIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import onError from '@lib/onError';
@@ -17,7 +16,7 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { SETTINGS } from 'src/tracking';
-import { Button, Card, Form, Input, useZodForm } from 'ui';
+import { Button, Card, Form, Input, Spinner, useZodForm } from 'ui';
 import getSignature from 'utils/getSignature';
 import getTokenImage from 'utils/getTokenImage';
 import { useContractWrite, useSignTypedData } from 'wagmi';
