+++ apps/web/src/components/Settings/Delete/index.tsx
@@ -1,6 +1,5 @@
 import MetaTags from '@components/Common/MetaTags';
 import UserProfile from '@components/Shared/UserProfile';
-import { Spinner } from '@components/UI/Spinner';
 import { WarningMessage } from '@components/UI/WarningMessage';
 import { useDisconnectXmtp } from '@components/utils/hooks/useXmtpClient';
 import { ExclamationIcon, TrashIcon } from '@heroicons/react/outline';
@@ -18,7 +17,7 @@ import toast from 'react-hot-toast';
 import Custom404 from 'src/pages/404';
 import { useAppPersistStore, useAppStore } from 'src/store/app';
 import { PAGEVIEW } from 'src/tracking';
-import { Button, Card, GridItemEight, GridItemFour, GridLayout, Modal } from 'ui';
+import { Button, Card, GridItemEight, GridItemFour, GridLayout, Modal, Spinner } from 'ui';
 import getSignature from 'utils/getSignature';
 import { useContractWrite, useDisconnect, useSignTypedData } from 'wagmi';
