+++ apps/web/src/components/Settings/Profile/index.tsx
@@ -1,5 +1,4 @@
 import MetaTags from '@components/Common/MetaTags';
-import { PageLoading } from '@components/UI/PageLoading';
 import { PhotographIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import { t } from '@lingui/macro';
@@ -13,7 +12,7 @@ import Custom404 from 'src/pages/404';
 import Custom500 from 'src/pages/500';
 import { useAppStore } from 'src/store/app';
 import { PAGEVIEW } from 'src/tracking';
-import { Card, GridItemEight, GridItemFour, GridLayout } from 'ui';
+import { Card, GridItemEight, GridItemFour, GridLayout, PageLoading } from 'ui';
 
 import SettingsSidebar from '../Sidebar';
 import NftPicture from './NftPicture';
