+++ apps/web/src/components/Shared/IndexStatus.tsx
@@ -1,4 +1,3 @@
-import { Spinner } from '@components/UI/Spinner';
 import { CheckCircleIcon } from '@heroicons/react/solid';
 import { Trans } from '@lingui/macro';
 import clsx from 'clsx';
@@ -6,6 +5,7 @@ import { POLYGONSCAN_URL } from 'data/constants';
 import { useHasTxHashBeenIndexedQuery } from 'lens';
 import type { FC } from 'react';
 import { useState } from 'react';
+import { Spinner } from 'ui';
 
 interface IndexStatusProps {
   type?: string;
