+++ apps/web/src/components/Shared/Modal/Report/index.tsx
@@ -1,4 +1,3 @@
-import { Spinner } from '@components/UI/Spinner';
 import { TextArea } from '@components/UI/TextArea';
 import { PencilAltIcon } from '@heroicons/react/outline';
 import { CheckCircleIcon } from '@heroicons/react/solid';
@@ -11,7 +10,7 @@ import type { FC } from 'react';
 import { useEffect, useState } from 'react';
 import { useGlobalModalStateStore } from 'src/store/modals';
 import { PAGEVIEW, PUBLICATION } from 'src/tracking';
-import { Button, EmptyState, ErrorMessage, Form, useZodForm } from 'ui';
+import { Button, EmptyState, ErrorMessage, Form, Spinner, useZodForm } from 'ui';
 import { object, string } from 'zod';
 
 import Reason from './Reason';
