+++ apps/web/src/components/Shared/Unfollow.tsx
@@ -1,4 +1,3 @@
-import { Spinner } from '@components/UI/Spinner';
 import { UserRemoveIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import onError from '@lib/onError';
@@ -15,7 +14,7 @@ import { useState } from 'react';
 import toast from 'react-hot-toast';
 import { useAppStore } from 'src/store/app';
 import { PROFILE } from 'src/tracking';
-import { Button } from 'ui';
+import { Button, Spinner } from 'ui';
 import getSignature from 'utils/getSignature';
 import { useSigner, useSignTypedData } from 'wagmi';
